+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # CHANGELOG
 
+## 0.5.13
+
+* docs: Add rationale.
+
 ## 0.5.12
 
 * test: Reorganize test namespaces.
 * fix: Fix path forming.
 * fix: Add type checking for variance in `TypeMapping` and `MappingEntry`.
 * refactor: Refactor `Transformer` for clarity.
+* refactor: Transformer is now lazy-loaded.
 
 ## 0.5.11